+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -9,7 +9,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.repository.CartRepository
 import woowacourse.shopping.data.localdb.ShoppingDB
 import woowacourse.shopping.ui.theme.AndroidshoppingTheme
