+++ app/src/main/java/woowacourse/shopping/ui/shopping/ShoppingActivity.kt
@@ -7,9 +7,9 @@ import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.data.MockProductRepository
-import woowacourse.shopping.data.RecentItemRepository
+import woowacourse.shopping.data.repository.CartRepository
+import woowacourse.shopping.data.repository.MockProductRepository
+import woowacourse.shopping.data.repository.RecentItemRepository
 import woowacourse.shopping.data.localdb.ShoppingDB
 import woowacourse.shopping.ui.cart.CartActivity
 import woowacourse.shopping.ui.detail.DetailActivity
