+++ app/src/test/java/woowacourse/shopping/data/CartRepositoryTest.kt
@@ -2,6 +2,7 @@ package woowacourse.shopping.data
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import woowacourse.shopping.data.repository.CartRepository
 import woowacourse.shopping.model.Money
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.ProductName
